// base/interfaces.ts
export type IDocumentReadyState = 'complete' | 'interactive' | 'loading';
// ...
export type IVisibilityState = 'hidden' | 'prerender' | 'visible';

export type IEventListenerOrEventListenerObject = IEventListener | IEventListenerObject;

export type IUnsupportedFeature = void;

export type ICanvasRenderingContext2D = IUnsupportedFeature;

export type IWebGLRenderingContext = IUnsupportedFeature;

export type IWindowProxy = IUnsupportedFeature;
// ...
export interface IGetRootNodeOptions {
  composed?: boolean;
}
// ...
export interface ICDATASection extends IText {
}
// ...
export interface IComment extends ICharacterData {
}
// ...
export interface IDocument extends INode, IDocumentOrShadowRoot, IGlobalEventHandlers, INonElementParentNode, IParentNode {
  readonly URL: string;
  readonly anchors: IHTMLCollection;
  readonly applets: IHTMLCollection;
  body: IHTMLElement | null;
  readonly characterSet: string;
  readonly charset: string;
  readonly compatMode: string;
  readonly contentType: string;
  cookie: string;
  readonly currentScript: IHTMLScriptElement | null;
  readonly defaultView: IWindowProxy | null;
  dir: string;
  readonly doctype: IDocumentType | null;
  readonly documentElement: IElement | null;
  readonly documentURI: string;
  readonly embeds: IHTMLCollection;
  readonly forms: IHTMLCollection;
  readonly head: IHTMLHeadElement | null;
  readonly hidden: boolean;
  readonly images: IHTMLCollection;
  readonly implementation: IDOMImplementation;
  readonly inputEncoding: string;
  readonly lastModified: string;
  readonly links: IHTMLCollection;
  readonly location: ILocation | null;
  onreadystatechange: ((this: IDocument, ev: IEvent) => any) | null;
  onvisibilitychange: ((this: IDocument, ev: IEvent) => any) | null;
  readonly plugins: IHTMLCollection;
  readonly readyState: IDocumentReadyState;
  readonly referrer: string;
  readonly scripts: IHTMLCollection;
  title: string;
  readonly visibilityState: IVisibilityState;
  adoptNode(node: INode): INode;
  captureEvents(): void;
  clear(): void;
  close(): void;
  createAttribute(localName: string): IAttr;
  createAttributeNS(namespace: string | null, qualifiedName: string): IAttr;
  createCDATASection(data: string): ICDATASection;
  createComment(data: string): IComment;
  createDocumentFragment(): IDocumentFragment;
  createElement(localName: string, options?: IElementCreationOptions): IElement;
  createElementNS(namespace: string | null, qualifiedName: string, options?: IElementCreationOptions): IElement;
  createEvent(eventInterface: string): IEvent;
  createNodeIterator(root: INode, whatToShow?: number, filter?: INodeFilter | null): INodeIterator;
  createProcessingInstruction(target: string, data: string): IProcessingInstruction;
  createRange(): IRange;
  createTextNode(data: string): IText;
  createTreeWalker(root: INode, whatToShow?: number, filter?: INodeFilter | null): ITreeWalker;
  getElementsByClassName(classNames: string): IHTMLCollection;
  getElementsByName(elementName: string): INodeList;
  getElementsByTagName(qualifiedName: string): IHTMLCollection;
  getElementsByTagNameNS(namespace: string | null, localName: string): IHTMLCollection;
  getSelection(): ISelection | null;
  hasFocus(): boolean;
  importNode(node: INode, deep?: boolean): INode;
  open(type?: string, replace?: string): IDocument;
  releaseEvents(): void;
  write(...text: string[]): void;
  writeln(...text: string[]): void;
  addEventListener<K extends keyof IDocumentEventMap>(type: K, listener: (this: IDocument, ev: IDocumentEventMap[K]) => any, options?: boolean | IAddEventListenerOptions): void;
  addEventListener(type: string, listener: IEventListenerOrEventListenerObject, options?: boolean | IAddEventListenerOptions): void;
  removeEventListener<K extends keyof IDocumentEventMap>(type: K, listener: (this: IDocument, ev: IDocumentEventMap[K]) => any, options?: boolean | IEventListenerOptions): void;
  removeEventListener(type: string, listener: IEventListenerOrEventListenerObject, options?: boolean | IEventListenerOptions): void;
}

export interface IDocumentFragment extends INode, INonElementParentNode, IParentNode {
}

export interface IDocumentOrShadowRoot {
  readonly activeElement: IElement | null;
}

export interface IDocumentType extends INode, IChildNode {
  readonly name: string;
  readonly publicId: string;
  readonly systemId: string;
}

export interface IElement extends INode, IChildNode, INonDocumentTypeChildNode, IParentNode, ISlotable {
  readonly attributes: INamedNodeMap;
  readonly classList: IDOMTokenList;
  className: string;
  readonly clientHeight: number;
  readonly clientLeft: number;
  readonly clientTop: number;
  readonly clientWidth: number;
  id: string;
  innerHTML: string;
  readonly localName: string;
  readonly namespaceURI: string | null;
  outerHTML: string;
  readonly prefix: string | null;
  readonly scrollHeight: number;
  scrollLeft: number;
  scrollTop: number;
  readonly scrollWidth: number;
  readonly shadowRoot: IShadowRoot | null;
  slot: string;
  readonly tagName: string;
  attachShadow(init: IShadowRootInit): IShadowRoot;
  closest(selectors: string): IElement | null;
  getAttribute(qualifiedName: string): string | null;
  getAttributeNS(namespace: string | null, localName: string): string | null;
  getAttributeNames(): Iterable<string>;
  getAttributeNode(qualifiedName: string): IAttr | null;
  getAttributeNodeNS(namespace: string | null, localName: string): IAttr | null;
  getBoundingClientRect(): IDOMRect;
  getClientRects(): IDOMRectList;
  getElementsByClassName(classNames: string): IHTMLCollection;
  getElementsByTagName(qualifiedName: string): IHTMLCollection;
  getElementsByTagNameNS(namespace: string | null, localName: string): IHTMLCollection;
  hasAttribute(qualifiedName: string): boolean;
  hasAttributeNS(namespace: string | null, localName: string): boolean;
  hasAttributes(): boolean;
  insertAdjacentElement(where: string, element: IElement): IElement | null;
  insertAdjacentHTML(position: string, text: string): void;
  insertAdjacentText(where: string, data: string): void;
  matches(selectors: string): boolean;
  removeAttribute(qualifiedName: string): void;
  removeAttributeNS(namespace: string | null, localName: string): void;
  removeAttributeNode(attr: IAttr): IAttr;
  setAttribute(qualifiedName: string, value: string): void;
  setAttributeNS(namespace: string | null, qualifiedName: string, value: string): void;
  setAttributeNode(attr: IAttr): IAttr | null;
  setAttributeNodeNS(attr: IAttr): IAttr | null;
  toggleAttribute(qualifiedName: string, force?: boolean): boolean;
  webkitMatchesSelector(selectors: string): boolean;
}

export interface IElementCSSInlineStyle {
  readonly style: ICSSStyleDeclaration;
}

export interface IElementContentEditable {
}

export interface IEvent {
  readonly AT_TARGET: number;
  readonly BUBBLING_PHASE: number;
  readonly CAPTURING_PHASE: number;
  readonly NONE: number;
  readonly bubbles: boolean;
  cancelBubble: boolean;
  readonly cancelable: boolean;
  readonly composed: boolean;
  readonly currentTarget: IEventTarget | null;
  readonly defaultPrevented: boolean;
  readonly eventPhase: number;
  readonly isTrusted: boolean;
  returnValue: boolean;
  readonly srcElement: IEventTarget | null;
  readonly target: IEventTarget | null;
  readonly timeStamp: number;
  readonly type: string;
  composedPath(): Iterable<IEventTarget>;
  initEvent(type: string, bubbles?: boolean, cancelable?: boolean): void;
  preventDefault(): void;
  stopImmediatePropagation(): void;
  stopPropagation(): void;
}

export interface IEventListenerObject {
  handleEvent(evt: IEvent): void;
}

export interface IEventTarget {
  dispatchEvent(event: IEvent): boolean;
}
// ...
export interface IHTMLCollection<T extends IElement = IElement> extends IHTMLCollectionBase {
  namedItem(name: string): T | null;
}
// ...
export interface INode extends IEventTarget {
  readonly ATTRIBUTE_NODE: number;
  readonly CDATA_SECTION_NODE: number;
  readonly COMMENT_NODE: number;
  readonly DOCUMENT_FRAGMENT_NODE: number;
  readonly DOCUMENT_NODE: number;
  readonly DOCUMENT_POSITION_CONTAINED_BY: number;
  readonly DOCUMENT_POSITION_CONTAINS: number;
  readonly DOCUMENT_POSITION_DISCONNECTED: number;
  readonly DOCUMENT_POSITION_FOLLOWING: number;
  readonly DOCUMENT_POSITION_IMPLEMENTATION_SPECIFIC: number;
  readonly DOCUMENT_POSITION_PRECEDING: number;
  readonly DOCUMENT_TYPE_NODE: number;
  readonly ELEMENT_NODE: number;
  readonly ENTITY_NODE: number;
  readonly ENTITY_REFERENCE_NODE: number;
  readonly NOTATION_NODE: number;
  readonly PROCESSING_INSTRUCTION_NODE: number;
  readonly TEXT_NODE: number;
  readonly baseURI: string;
  readonly childNodes: INodeList;
  readonly firstChild: INode | null;
  readonly isConnected: boolean;
  readonly lastChild: INode | null;
  readonly nextSibling: INode | null;
  readonly nodeName: string;
  readonly nodeType: number;
  nodeValue: string | null;
  readonly ownerDocument: IDocument | null;
  readonly parentElement: IElement | null;
  readonly parentNode: INode | null;
  readonly previousSibling: INode | null;
  textContent: string | null;
  appendChild(node: INode): INode;
  cloneNode(deep?: boolean): INode;
  compareDocumentPosition(other: INode): number;
  contains(other: INode | null): boolean;
  getRootNode(options?: IGetRootNodeOptions): INode;
  hasChildNodes(): boolean;
  insertBefore(node: INode, child: INode | null): INode;
  isDefaultNamespace(namespace: string | null): boolean;
  isEqualNode(otherNode: INode | null): boolean;
  isSameNode(otherNode: INode | null): boolean;
  lookupNamespaceURI(prefix: string | null): string | null;
  lookupPrefix(namespace: string | null): string | null;
  normalize(): void;
  removeChild(child: INode): INode;
  replaceChild(node: INode, child: INode): INode;
}

export interface INodeIterator {
  readonly filter: INodeFilter | null;
  readonly pointerBeforeReferenceNode: boolean;
  readonly referenceNode: INode;
  readonly root: INode;
  readonly whatToShow: number;
  detach(): void;
  nextNode(): INode | null;
  previousNode(): INode | null;
}

export interface INodeList<T extends INode = INode> {
  readonly length: number;
  item(index: number): T | null;
  forEach(callbackfn: (value: INode, key: number, parent: INodeList<T>) => void, thisArg?: any): void;
  [index: number]: T;
}

export interface INonDocumentTypeChildNode {
  readonly nextElementSibling: IElement | null;
  readonly previousElementSibling: IElement | null;
}

export interface INonElementParentNode {
  getElementById(elementId: string): IElement | null;
}

export interface IParentNode {
  readonly childElementCount: number;
  readonly children: IHTMLCollection;
  readonly firstElementChild: IElement | null;
  readonly lastElementChild: IElement | null;
  append(...nodes: (INode | string)[]): void;
  prepend(...nodes: (INode | string)[]): void;
  querySelector(selectors: string): IElement | null;
  querySelectorAll(selectors: string): INodeList;
}
// ...
export interface IHTMLElement extends IElement, IElementCSSInlineStyle, IElementContentEditable, IGlobalEventHandlers, IHTMLOrSVGElement {
  accessKey: string;
  dir: string;
  draggable: boolean;
  hidden: boolean;
  lang: string;
  readonly offsetHeight: number;
  readonly offsetLeft: number;
  readonly offsetParent: IElement | null;
  readonly offsetTop: number;
  readonly offsetWidth: number;
  title: string;
  translate: boolean;
  click(): void;
  addEventListener<K extends keyof IGlobalEventHandlersEventMap>(type: K, listener: (this: IHTMLElement, ev: IGlobalEventHandlersEventMap[K]) => any, options?: boolean | IAddEventListenerOptions): void;
  addEventListener(type: string, listener: IEventListenerOrEventListenerObject, options?: boolean | IAddEventListenerOptions): void;
  removeEventListener<K extends keyof IGlobalEventHandlersEventMap>(type: K, listener: (this: IHTMLElement, ev: IGlobalEventHandlersEventMap[K]) => any, options?: boolean | IEventListenerOptions): void;
  removeEventListener(type: string, listener: IEventListenerOrEventListenerObject, options?: boolean | IEventListenerOptions): void;
}
// ...
export interface IHTMLScriptElement extends IHTMLElement {
  charset: string;
  crossOrigin: string | null;
  defer: boolean;
  event: string;
  htmlFor: string;
  src: string;
  text: string;
  type: string;
  addEventListener<K extends keyof IGlobalEventHandlersEventMap>(type: K, listener: (this: IHTMLScriptElement, ev: IGlobalEventHandlersEventMap[K]) => any, options?: boolean | IAddEventListenerOptions): void;
  addEventListener(type: string, listener: IEventListenerOrEventListenerObject, options?: boolean | IAddEventListenerOptions): void;
  removeEventListener<K extends keyof IGlobalEventHandlersEventMap>(type: K, listener: (this: IHTMLScriptElement, ev: IGlobalEventHandlersEventMap[K]) => any, options?: boolean | IEventListenerOptions): void;
  removeEventListener(type: string, listener: IEventListenerOrEventListenerObject, options?: boolean | IEventListenerOptions): void;
}