// src/api/Element.d.ts
export default class Element extends GeneratedElement implements IElement {
    constructor();
    get classList(): IDOMTokenList;
    get className(): string;
    get id(): string;
    get outerHTML(): string;
    get innerHTML(): string;
    closest(selectors: string): IElement | null;
    getAttribute(qualifiedName: string): string | null;
    getAttributeNS(namespace: string | null, localName: string): string | null;
    getAttributeNode(name: string): IAttr | null;
    getAttributeNodeNS(namespaceURI: string, localName: string): IAttr | null;
    getElementsByClassName(classNames: string): IHTMLCollection;
    getElementsByTagName(qualifiedName: string): IHTMLCollection<IElement | any>;
    getElementsByTagNameNS(namespaceURI: string, localName: string): IHTMLCollection<IHTMLElement | ISVGElement>;
    hasAttribute(qualifiedName: string): boolean;
    hasAttributeNS(namespace: string | null, localName: string): boolean;
    hasAttributes(): boolean;
    removeAttribute(qualifiedName: string): void;
    removeAttributeNS(namespace: string | null, localName: string): void;
    removeAttributeNode(attr: IAttr): IAttr;
    setAttribute(qualifiedName: string, value: string): void;
    setAttributeNS(namespace: string | null, qualifiedName: string, value: string): void;
    setAttributeNode(attr: IAttr): IAttr | null;
    setAttributeNodeNS(attr: IAttr): IAttr | null;
}